wip: first calculations

+++ database/migrations/2025_01_21_232332_create_purchases_items_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->id();
             $table->biginteger("grocery_item_id")->unsigned();
             $table->integer("place_id")->unsigned()->nullable();
-            $table->integer("amount")->unsigned()->nullable();
+            $table->decimal("amount", 7, 2)->unsigned()->nullable();
             $table->timestamps();
             $table->foreign("grocery_item_id")->references("id")->on("groceries_items");
         });
